Move more legacy code, fix more imports
I said that this was like untangling rope, but it's more like brushing a badly groomed dog.
