+++ HARK/ConsumptionSaving/ConsLabeledModel.py
@@ -14,11 +14,11 @@ from HARK.ConsumptionSaving.ConsPortfolioModel import (
     init_portfolio,
 )
 from HARK.ConsumptionSaving.ConsRiskyAssetModel import (
-    FixedPortfolioShareRiskyAssetConsumerType,
     RiskyAssetConsumerType,
     init_risky_asset,
     init_risky_share_fixed,
 )
+from HARK.ConsumptionSaving.FixedPortfolioShareModel import FixedPortfolioShareRiskyAssetConsumerType
 from HARK.ConsumptionSaving.LegacyOOsolvers import ConsIndShockSetup
 from HARK.core import make_one_period_oo_solver
 from HARK.distribution import DiscreteDistributionLabeled
